requirements.txt
PyQt6==6.9.0
PyQt6-WebEngine==6.9.0
send2trash==1.8.0
Pillow==11.3.0
httpx==0.27.2
openai==1.44.0
requests==2.32.5
torch==2.8.0
sentence-transformers==5.1.1
imagehash==4.3.2
python-xmp-toolkit==2.0.2
PyYAML==6.0.3
huggingface_hub[hf_xet]>=0.34.0,<1.0
transformers>=4.44,<5.0
accelerate>=1.0.0
ultralytics>=8.0.0,<9.0.0
scikit-image==0.22.0